requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
google==3.0.0
google-search==1.1.1
requests==2.28.2
Pillow==9.4.0
selenium==4.7.2
twilio==7.16.1
